Extract GravatarSnackbarHost to :design module
