+++ homeUi/build.gradle.kts
@@ -11,6 +11,7 @@ android {
 
 dependencies {
     implementation(project(":userComponent"))
+    implementation(project(":design"))
 
     implementation(libs.androidx.core.ktx)
     implementation(libs.androidx.material3)
